Refactoring IReportScriptDataProvider and review comments
