+++ CDP4Reporting.Tests/ViewModels/ReportDesignerViewModelTextFixture.cs
@@ -107,13 +107,13 @@ namespace CDP4Reporting.Tests.ViewModels
         private const string DATASOURCE_CODE_WITH_PARAMS = @"namespace CDP4Reporting
         {
             using CDP4Reporting.DataCollection;
-	        using CDP4Reporting.Parameters;
-	        using System.Collections.Generic;
+            using CDP4Reporting.Parameters;
+            using System.Collections.Generic;
 
-	        public class TestReportingParameters: ReportingParameters
-	        {
-		        public override IEnumerable<IReportingParameter> CreateParameters(object dataObject)
-		        {
+            public class TestReportingParameters: ReportingParameters
+            {
+                public override IEnumerable<IReportingParameter> CreateParameters(object dataObject)
+                {
                     List<IReportingParameter> paramsList = new List<IReportingParameter>();
                     paramsList.Add(new ReportingParameter(""param1"", typeof(int), 0));
                     paramsList.Add(new ReportingParameter(""param2"", typeof(string), string.Empty));
@@ -123,8 +123,8 @@ namespace CDP4Reporting.Tests.ViewModels
                     paramsList.Add(new ReportingParameter(""param3"", typeof(string), string.Empty));
                     paramsList[3].Visible = false;
 
-			        return paramsList;
-		        }
+                    return paramsList;
+                }
             }
             public class TestDataSource : DataCollector
             {
