+++ CDP4Reporting/DataCollection/IDataCollectorParameter.cs
@@ -1,5 +1,5 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="DataCollectorParameterColumn.cs" company="RHEA System S.A.">
+// <copyright file="IDataCollectorParameter.cs" company="RHEA System S.A.">
 //    Copyright (c) 2015-2020 RHEA System S.A.
 //
 //    Author: Sam Gerené, Alex Vorobiev, Alexander van Delft, Cozmin Velciu, Adrian Chivu
