+++ CDP4Reporting/Parameters/ReportingParameters.cs
@@ -28,55 +28,13 @@ namespace CDP4Reporting.Parameters
     using System.Collections.Generic;
     using System.Linq;
 
-    using CDP4Common.EngineeringModelData;
-    using CDP4Common.SiteDirectoryData;
-
-    using CDP4Dal;
+    using CDP4Reporting.DataCollection;
 
     /// <summary>
     /// A class that is used to build Report Parameters and optional a specific filter string at the report level. 
     /// </summary>
-    public abstract class ReportingParameters : IReportingParameters
+    public abstract class ReportingParameters : ReportScriptDataProvider, IReportingParameters
     {
-        /// <summary>
-        /// Gets the <see cref="Iteration"/>
-        /// </summary>
-        public Iteration Iteration { get; private set; }
-
-        /// <summary>
-        /// Gets the <see cref="ISession"/>
-        /// </summary>
-        public ISession Session { get; private set; }
-
-        /// <summary>
-        /// Gets the <see cref="DomainOfExpertise"/>
-        /// </summary>
-        public DomainOfExpertise DomainOfExpertise { get; private set; }
-
-        /// <summary>
-        /// All currently open <see cref="ReferenceDataLibrary"/>s in this <see cref="IReportingParameters.Session"/>
-        /// </summary>
-        public IEnumerable<ReferenceDataLibrary> OpenReferenceDataLibraries { get; private set; }
-
-        /// <summary>
-        /// The current <see cref="SiteDirectory"/>s in this <see cref="IReportingParameters.Session"/>
-        /// </summary>
-        public SiteDirectory SiteDirectory { get; private set; }
-
-        /// <summary>
-        /// Initializes this DataCollector 
-        /// </summary>
-        /// <param name="iteration"></param>
-        /// <param name="session"></param>
-        public void Initialize(Iteration iteration, ISession session)
-        {
-            this.Iteration = iteration;
-            this.Session = session;
-            this.DomainOfExpertise = session.QueryCurrentDomainOfExpertise();
-            this.OpenReferenceDataLibraries = session.OpenReferenceDataLibraries;
-            this.SiteDirectory = session.RetrieveSiteDirectory();
-        }
-
         /// <summary>
         /// Creates a list of report reporting parameter that should dynamically be added to the 
         /// Report Designer's report parameter list.
